// src/app/shared/chart/models/ChartTypeEnum.ts
export enum ChartTypeEnum {
    VERTICAL_BAR = "VERTICAL_BAR",
    PIVOT_TABLE = "PIVOT_TABLE",
    HORIZONTAL_BAR = "HORIZONTAL_BAR",
    LINE = "line",
    STATUS_METER = "STATUS_METER",
    PIE = "PIE",
    BUBBLE = "BUBBLE",
    STACKED_VERTICAL_BAR = "vertical",
    TREE_MAP = "TREE_MAP",
    AREA = "area",
}